requirements.txt
python-dotenv
streamlit
groq
duckduckgo-search
wikipedia
arxiv
requests